requirements.txt
# Requirements for MLflow modular project with RFDETR
mlflow>=2.0.0
numpy>=1.20.0
pandas>=1.3.0
scikit-learn>=1.0.0
matplotlib>=3.4.0
typings>=3.7.4
joblib>=1.1.0
seaborn>=0.11.0
tqdm>=4.62.0
# RFDETR dependencies
rfdetr>=0.1.0
supervision>=0.6.0
Pillow>=9.0.0
torch>=1.10.0
torchvision>=0.11.0
requests>=2.27.0